pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sar"
version = "0.3.0"
description = "Security Application Registry (SAR)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0-or-later"}
dependencies = [
  "fastapi",
  "uvicorn",
  "pandas",
  "openpyxl",
  "jinja2",
  "python-multipart",
  "docxtpl",
  "pyyaml",
  "itsdangerous",
  "argon2-cffi"
]


[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "httpx",
  "freezegun"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[project.urls]
Homepage = "https://example.invalid/sar"

[project.scripts]
sar = "sar.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sar = [
  "templates/*.html",
  "static/**/*",
  "report_templates/*.docx",
  "report_templates/*.html.j2"
]
